Moved fetchComponents function in build order
