+++ deployment_specific/pb/ds_components/side-pane/side-pane.html
@@ -170,12 +170,12 @@
             </a>
           </td>
         </tr>
-          <tr class="logos">
-            <td colspan="2" align="center">
-              <a href="https://www.zurich.co.id/id-id/kampanye-kami/zurich-alert-the-alert-you-really-need">
-                <img class="partnerLogo" src="assets/graphics/zAlert_logo.png">
-              </a>
-            </td>
+        <tr class="logos">
+          <td colspan="2" align="center">
+            <a href="https://www.zurich.co.id/id-id/kampanye-kami/zurich-alert-the-alert-you-really-need">
+              <img class="partnerLogo" src="assets/graphics/zAlert_logo.png">
+            </a>
+          </td>
           <td colspan="2" align="center">
             <a href="https://twitter.com">
               <img class="partnerLogo" src="assets/graphics/twitter-logo.png">
@@ -188,7 +188,7 @@
               <img class="partnerLogo" src="assets/graphics/mapbox-logo.png">
             </a>
           </td>
-      </tr>
+        </tr>
       </table>
 
       <p innerhtml.bind="locale.info_content.p1"></p>
@@ -199,7 +199,7 @@
 
       <p id="mapbox">
         ${locale.info_content.mapCredits} ©
-        <a href="https://www.mapbox.com/about/maps/" target="_blank">Mapbox</a> © <a     href="http://www.openstreetmap.org/about"   target="_blank">OpenStreetMap</a>.
+        <a href="https://www.mapbox.com/about/maps/" target="_blank">Mapbox</a> © <a href="http://www.openstreetmap.org/about" target="_blank">OpenStreetMap</a>.
         ${locale.info_content.mapImproveLink} <a href="https://www.mapbox.com/map-feedback/" target="_blank">Mapbox</a>.
       </p>
       <p id="browserstack" innerhtml.bind="locale.info_content.browserStack">
